requirements.txt
django==3.2.15
pytz==2022.7
djangorestframework==3.14.0
Markdown==3.4.1
# pandas==1.5.2
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
ipython==8.8.0
python-dotenv==0.21.0
# black==
# isort==